Fix tests for latest aktualizr
The tests got broken between 2020.2 and 2020.3 because of a change to
how non-standard packagemanager configuration was handled. As a result,
we can't directly manipulate what DockerAppPackageManager is looking at
any more. We have to re-construct the object after each change to its
configuration. Not a big deal for a unit test

Signed-off-by: Andy Doan <andy@foundries.io>