Get rid of of the ostree mock
Effectively the mock was required for getting a checksum of the
currently installed ostree revision/commit since the ostree package
manager fails to get it if a host is not booted into non the ostree-based
rootfs. But, we can get a checksum of the deployment that is
present in the ostree repo by simply listing the available deployments
in the repo and getting its commit checksum.

Signed-off-by: Mike Sul <mike.sul@foundries.io>