Don't (re-)start apps containers if reboot
This change makes compose apps's containers not to (re-)start
just after their update and (re-)creation if we are about to reboot a
system in order to apply an ostree-based rootfs update.
The updated apps will be started just after a system boot if aklite runs in a daemon mode.
If it is at a non-daemon mode then the updated apps will be started on a first run
of aklite or they can be started manually.

Signed-off-by: Mike Sul <mike.sul@foundries.io>