Fix a race condition in the event reporting
It turned out that there are two instances of ReportQueue running in the
context of aktualizr-lite. One instance is owned by LiteClient the
second one by SotaUptaneClient. Each instance runs a thread that fetches
events from the DB and sends them to the backend, this procedure is not
synchronized what causes the race condition with side effects like
sending the same event twice.
The key question here is why do we need both clients LiteClient and SotaUptaneClient?
Taking into account that SotaUptaneClient is tuned for the Uptane
use-case it doesn't make sense to use it as it requires hacking around in order
to eliminate the Uptane specific logic that is not required in the TUF
case as well as the logic around Secondary support. Instead we can just
reuse the libaktualizr's mechanisms in LiteClient in the way required to fulfil our
specific use-cases. Doing so helps us to fix the immediate issue with
two ReportQueue instances as well as achieve the strategic goal with
minimizing dependency on and vulnerability to changes in libaktualizr.

Signed-off-by: Mike Sul <mike.sul@foundries.io>