Fix sending the installation applied event
The event EcuInstallationApplied should be sent to the backend if an
ostree update is successfully fetched and installed (checkout but system
is not booted into it yet). The issue was that a system reboot was
invoked in the context of running LiteClient without doing
its tearing down prior to it. As result, the event was either persisted in
the DB and sent after reboot or sent actually before reboot if
ReportQueue's thread managed to fetch and sent it before reboot
or simply lost due to the race condition caused by two active ReportQueue
instances and on-going reboot.
This fix just simply lets LiteClient to deinitialize/teardown its context
properly including flushing ReportQueue before triggering a system
reboot.

Signed-off-by: Mike Sul <mike.sul@foundries.io>