Fix comparision of a new and installed Targets
A list of apps specified in the sota.toml configuration
as well as what is actually located on a file system should be
taken into account during comparision of the new and currently
installed Targets. Otherwise, aklite will keep trying to fetch and
install Target that is actually installed or never re-install an app
that has been removed from a file system.
Use cases:
 - a list of apps in the config is subset of Target's list of apps
 - an app has been removed from a file system

Signed-off-by: Mike Sul <mike.sul@foundries.io>