fixup! Update Apps if a new Target after restart
Make sure that Apps to be updated are checked just after aklite restart
if there is a new Target. It does not suffice to verify the app list
in the config and if currently installed Target's Apps are installed
and running. A new check has to happen if a new Target is discovered.

Signed-off-by: Mike Sul <mike.sul@foundries.io>