aklite: check for new Target before checking apps
At first, check if there is a new Target, and if so then proceed with new
Target update. If there is no a new Target then check if the apps list
changed in the config (first run) and if all current Target' apps are installed and running.
There is no any reason to perform current Target' apps check and sync if
there is a new Target since the new Target can have a different list of apps.

Signed-off-by: Mike Sul <mike.sul@foundries.io>