aklite: allow update to previously installed Target
Currently aklite doesn't allow to install any Target that it tried to
install previously. Even if Target installation was successful it treats
it as a "bad" Target that caused rollback and doesn't allow installing
it.
The given change checks Target to be installed against Targets that
aklite tried to install but never installed them except pending one,
what can happen only in the case of rollback.

Signed-off-by: Mike Sul <mike.sul@foundries.io>