aklite: optimize retrieval of current Target
There are many spots in the source code where currently installed and
running Target is needed and each time getCurrent is called it goes down
to the DB, checks which of the commits is currently deployed and in the
future we actually need to verify which of the apps are actually
installed and running. On the other hand, we really need to retrieve and
determine what currently is installed and running just once per an
update cycle in a daemon mode and once if it's non-daemon mode.
The ultimate goal is to make getCurrent() return Target that list only
those elements/apps that are really installed and running not just part
of TUF Target.

Signed-off-by: Mike Sul <mike.sul@foundries.io>