composeinfo: parse yaml file
This commit adds ComposeInfo parser to read all the services
described in a yaml file as well as retrieving the corresponding image
paths for each service from the file.

It relies on https://github.com/pantoniou/libfyaml being installed in
the target in order to perform the yaml to json conversion.

The corresponding functional tests are also part of this commit.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>