apps: check if App is installed
Initial version of verification whether App is installed.
Since it's up to docker-compose/docker to decide if App's container
should be running or not and manage their state we/aklite still need
some criterion to decide whether a given App needs to (re-)installed.
The key use-cases:
- adding App to compose_apps list (shortlist)
- removing compose_apps/shortlist at all, all Target apps should be
installed
- unintentional removal/modification of App content located on a file system
- rollback on boot after a new Target's Apps has been installed but not
running. The previous Target's Apps needs to be reinstalled in this
case.

Signed-off-by: Mike Sul <mike.sul@foundries.io>