api: Allow check-ins to find targets for secondary ecus
There are two ways a user can do secondary ECUs. The recommended path
is with a single target where the ostree hash of the secondary ecu
is embedded in the primary's custom data.

Some users have a Target dedicated to the secondary ECU. This change
allows them to look up that target by hwid.

Signed-off-by: Andy Doan <andy@foundries.io>