api: Introduce InstallContext
Trying to capture the correlation-id and client and uptane target within
the DownloadResult via the instal function pointer starts to break down
with the introduction of secondary ecus. They will need access to this
same context data to issue things like update events. This change
creates a new InstallContext abstract class that can be returned and
used for normal installs as well as future secondary ecu updates.

Signed-off-by: Andy Doan <andy@foundries.io>