apps: verify App at the Target verification phase
Once Target is successfully downloaded a liteclient's client invokes
the Target verification. For some reason it's never been used by
ComposeAppEngine before and the verification happens at the end of fetch
step.
The given PR adds the Target verification support to ComposeAppManager
and both ComposeApp and RestorableApp engines.
Since "verification" at the end of "fetch" phase became a feature for
ComposeApp, there is no much reason to remove it.
In the case of RestorableAppEngine it will help to distinguish between
actuall/true download failure and invalid Target. In the first case,
aklite should keep trying to download a new Target, while in the case of
an invalid Target it doesn't make much sense in the update retrying.

Signed-off-by: Mike Sul <mike.sul@foundries.io>