aklite: store rollback versions in LiteClient context
If we want to detect and set rolbacks within an aklite update cycle,
not just after a system reboot, then we need to be able to maintain
the list of "bad" versions in the aklite context.
Otherwise, aklite clients will have to maintain/update the "bad" version
list for each rollback use-case what leads to code copy-paste in each
client.

Signed-off-by: Mike Sul <mike.sul@foundries.io>