apps: check expected size of a layer manifest
We do not know in advance a size of an App's index/root manifest,
hence we just assume that it shouldn't be bigger than certain
predefined value (16KB).
But, we know a layer manifest size since it is defined in the root
level manifest (Merkle tree).
Therefore, we must ensure that a downloaded layer manifest's size
matches the size specified in the root manifest. Also, we must sure that
data being downloaded don't exceed the given manifest's size,
libcurl does that we just pass a corresponding param down
to it via the http client.

Signed-off-by: Mike Sul <mike.sul@foundries.io>