apps: check both skopeo and docker store before update
Check whether there is sufficient amount of available storage in both
the skopeo and docker store before an App update.

A storage size, required to accommodate a new App version, is calculated
precisely for the skopeo store (tar.gzip format) and approximately
for the docker store.
The approximation is based on an average gzip compression ratio in
the case of rootfs diffs/image layers.

Signed-off-by: Mike Sul <mike.sul@foundries.io>