aklite: fix comparision with `current` while checking for rollback
Current Target should be omitted from checking whether Target is
`known locally` (bad Target). The given change fixes the comparision
of a given Target with Current Target. It's not enough to compare just
hashes since Current and a specified Targets can have the same ostree
hash while still be different because of difference in Apps.
Since each Target has an unique `filename` (aka name) it's suffice to
compare only Targets' filenames.

Added a test for this specific case - failure in the case of
just App update so a rollback to the current version should happen,
which is actually just App sync for the current Target.

Signed-off-by: Mike Sul <mike.sul@foundries.io>