aktualizr: allow updating meta if the same ver
Allow updating metadata if local and remote versions are the same
but they actually differ.
We need it for the use-case of switching tags on a device what may lead
to a need to update metadata of the same version.
Each tag has its own set of metadata (except root) so metadata
of different tags can have the same version but they actually differ
because they refer to different set of targets.

Signed-off-by: Mike Sul <mike.sul@foundries.io>