custom-client: check installer before usage
Make sure that a returned installer is actually initialized before
using/dereferencing it.
It might happen in the case when the checkin was successful and new
target has been found but for some reason the new target metadata were
not stored in the DB, so the following request to obtain latest target
silently fails.

Signed-off-by: Mike Sul <mike.sul@foundries.io>