aklite: enforce ostree update if hashes differ
Enforce an Target/ostree update if current and the latest targets
have different ostree hashes even if their versions match.

This is needed for the use-case when the same OE CI run is re-run one
or more times and each time a new ostree and target are pushed to the
backend.

Signed-off-by: Mike Sul <mike.sul@foundries.io>