tests: Run tests as `testuser` in container
Running tests in the container as a root user leads to odd issues
related to removing of ostree objects. For some reason ostree repo
ends up with `i` (immutable) attribute on a file system and cannot
be removed by test.
In any case the tests should be run as a non-root user. The root user
is needed only for running `dockerd` at the container start.

Signed-off-by: Mike Sul <mike.sul@foundries.io>