tests: Add test for the boot flag mgmt
Add a test suite to test the boot flag management for each type of the
bootloader (aka rollback mode).

The test suite defines a parameterized test that initializes different
type of boot flag mgmt and different type of package manager.
The flags are checked in the base test client (liteclienttest.cc).

The given test suite can be used as a starting point for more advanced
tests of the boot flag management, e.g. check the flag values in the
case of rollback and unexpected device reboot.

Signed-off-by: Mike Sul <mike.sul@foundries.io>