offline: Fix the content based App shortlisting
It turned out that content-based App shortlisting was implemented in the
first place but there was a bug in the algorithm.
The algorithm makes sure that:

1. All provided Apps are listed in Target.
2. Not provided Apps are removed (shortlisted) from the resultant
   Target that is passed to the offline update client as an input.

The issue was in the "early" exit from the shortlisting loops so not all
Target apps where checked. The early exit happened when the algo
checked that all provided Apps are present in the Target App list.

Signed-off-by: Mike Sul <mike.sul@foundries.io>