client: Don't mark Target as pending if it is not
Mark Target being installed as "pending" only if it's really pending,
specifically, if its ostree hash has been deployed and a device will try
to boot on it on the next boot.
Aklite may try to install Target and exit earlier with the
"need completion" status if there is ongoing boot firmware update.
If it happens, then Target should not be marked as "pending".

Signed-off-by: Mike Sul <mike.sul@foundries.io>