offline: Just finalize installation to complete rollback
If there is a rollback during  "just App" update then a device is
already booted on a the correct ostree/rootfs version, so we need just
finalize rollback/update to the previous version since dockerd was
restarted.
If there is a rollback during "ostree + just App" update then we do a
regular install of the rollback/previous target, reboot, and then do
finalization.

Signed-off-by: Mike Sul <mike.sul@foundries.io>