appengine: Allow to configure concurrent pull numb
Override the default value of the skopeo's maximum parallel/concurrent
pulls if the env. var. `SKOPEO_MAX_PARALLEL_PULLS` is set to a value in
the allowed range [1, 10].
If `SKOPEO_MAX_PARALLEL_PULLS` is not defined, then `skopeo` will apply
the default value which is `3` for LmP and `6` for other use-cases.

Signed-off-by: Mike Sul <mike.sul@foundries.io>