restorableapps: Cleanup skopeo's tmp files
Remove tmp files created by skopeo during image layer pulling that are
not cleaned up if `skopeo copy` is interrupted.

The cleanup function is called only at the aklite initialization since the
tmp files are not removed by skopeo only if an overall process is
interrupted/killed/device is rebooted.

Signed-off-by: Mike Sul <mike.sul@foundries.io>