rootfsmanager: Use ostree repo path to check update size
Use a path to the system's ostree repo (`/ostree/repo`) as a path for
which to get afile system (volume) statistic. Using the sysroot path
`/sysroot` doesn't work because the libaktualizr set it to an empty
string by default, not to `sysroot` as expected.
Also, libostree uses the repo path in the call to get FS stats, so this
change makes aklite in sync with libostree in terms of checking if there
is enough free storage before an update.

Signed-off-by: Mike Sul <mike.sul@foundries.io>