stat: Introduce logic to obtain storage usage
Add functionality to get storage usage statistic and represent it
in a nice and convenient way. It allows us using the same storage
stat function across many use-cases.

The functionality includes a way to identify storage reservation type.

We have three knobs to control storage reservation:
1. The knob/param to set storage reservation for apps. The same value
   for both the skopeo and docker storage even if they are located on
   different volumes.
2. The knob to set `min-free-space-percent` parameter of the sysroot's
   ostree repo. This value is used by libostree during committing/adding
   an object to a repo. The check occurs per each object, just before
   adding it.
3. The knob to set the reserved space value that is used before pulling
   an ostree commit if static delta statistic is available.

Therefore, we need a way to identify these three types of knobs. For
example, it's useful to see in the storage usage statistic which of the
knobs is applied in a given case.

Signed-off-by: Mike Sul <mike.sul@foundries.io>