tests: Calculate exact expected numbers
A size of a new ostree commit may differ on different hosts due to some
randomness in some of additional files added to the sysroot filesystem.
Therefore, the delta size between a current and a new ostree commits is
equal <requested block number * block size> + <random size less than one
block>.
Therefore, to make the tests pass on any host we need to determine the
delta size exactly and check that the corresponding numbers are present
in the resultant update events.

Signed-off-by: Mike Sul <mike.sul@foundries.io>