rootfsmanager: Do pre-pull check correctly
There are two knobs to configure the same thing - a reserved amount of
storage capacity. One is used by libostree during ostree object
committing to a repo, another one is intended for the delta-based pull
use-case.
If the "ostree" knob reserves more storage than the "delta" knob then
it makes sense to take into account the "ostree" reservation during the
pre-pull update size verification, not the "delta" one.
Otherwise, libostree will yield an error during ostree pull once the
storage usage reaches the reserved threshold. So, no point in starting
an ostree pull if it will fail.
Ideally, we need to get rid of the "delta" know and use just the
"ostree" one.

Signed-off-by: Mike Sul <mike.sul@foundries.io>