apps: Assume stateless app is pulled when running it
It might happen that the compose app state file is missing when a call
to run it is invoked. This should not happen under normal circumstances
since the previous `fetch` call makes sure the state file is created and
set.
However, in the case of switching from the restorable app type to
compose one or manual app pull (`docker-compose pull/up` the app's
compose dir ends up with no state file at the moment of its running.
So, this fix assumed that the app is pulled if `run` is called and
enforces the app state file setting to the `pulled` state.
Even if the app compose content is missing the run command will pull it.

Signed-off-by: Mike Sul <mike.sul@foundries.io>