docker: client: Add method to load image
Add a new method to the docker client to load image into the docker
store. Effectively, the method makes `HTTP POST /images/load` request to
the docker daemon. Prior to it, it generates an in-memory TAR archive
that contains the image load manifest and named as `manifest.json`; this
is what the handler expects.

Signed-off-by: Mike Sul <mike.sul@foundries.io>