apps: Only load and never pull images
Don't fallback to pulling images from Registry by `docker compose pull`
command in the load of App images fails.
If the App imaged loading fails then treat it as `ImagePullFailure`
failure, so the Target is not marked as failing. Then, the image loading
will be retried during the next update cycle.
It makes sense to do so because the image load process may fail due to
some local unexpected issues, e.g. the docker daemon is down.

Signed-off-by: Mike Sul <mike.sul@foundries.io>