tests: Make offline tests environment agnostic
Make the offline update tests agnostic to the test environment.
Specifically, to the `free/capacity` ratio of the storage volume of the
VM/container/host the tests run on.

The recent changes and fixes to the lite client improved the update size
check. Most of the tests (online case) were switched to using the mock
function to get storage statistic, however the offline update part was
forgotten. So, this change fixes it.

Signed-off-by: Mike Sul <mike.sul@foundries.io>