offline: Check Target ostree commit presence correctly
It turned out that the current approach to finding ostree commits in the
provided ostree repo doesn't work correctly if multiple commits of the
same branch/ref is present in the repo. The reason is simple - a ref can
refence only single commit.

So, instead of parsing the ostree repo directory and trying to find
commit files there, let's check whether a given Target commit hash is
present in the repo through the libostree API.

Also, it makes sure that only those commits that are referenced from the
TUF Targets meta are "found" and considered for an update.

Signed-off-by: Mike Sul <mike.sul@foundries.io>