offline: Shortlist Target Apps correctly
The current logic that performs so-called content-based shortlisting is
not quite correct for the case when there are apps for two or more
Targets (if there is at least one App that is not part of a given target
then all target apps are shortlisted).
Also, this logic is too complex while it can be much simpler.
Therefore, this change fixes both issues, it:

1. Allows including Apps of more than one target to the offline update
   source directory.
2. Makes the content-based Target shortlisting simpler - it simply
   checks whether App content is available in the source directory
   for each Target App.

Signed-off-by: Mike Sul <mike.sul@foundries.io>