offline: Restructure Target filtering and finding
Refactor the target filtering, sorting, and finding functionality by
dividing it into separate functions, each encapsulating a specific
functional piece. This approach enhances modularity and maintainability.

This is needed for following commit where the functional blocks will be
reused in a new offline update sub-command.

Add test to verify the case when there are more than one target is
available.

Signed-off-by: Mike Sul <mike.sul@foundries.io>