offline: Fix --log-level option
The command line  option for log level in aklite-offline does not match
the one used in aktualizr ("log-level" vs "loglevel"). Libaktualizr sets
the log level regardless of the running application, and assumes a
"loglevel" option is set in the program options map.
This patch adds a "loglevel" entry to the map, copying the "log-level"
value.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>