offline: Require all enabled apps for offline update
It's not safe/secure to do "content-based" shortlisting as it can lead
to installing of unwanted target or removing one or more apps that are
actually required.

The new approach requires that all apps that are enabled and listed in
the target are present in the update source bundle (intersection of
`pacman.compose_app` and `target.apps`).

This guarantee that a device is updated with expected ostree and apps
(controlled via target json and device configuration).
It may happen that there are more than one target that has the same
ostree and the given app intersection/list. However, it is not an issue
because the update content is exactly the same across those targets. The
only drawback is that the offline updater can print not expected target
name in the logs and the current target name may be not what a user
expected. But again, the update content, ostree  version and app list
and their version is exactly what a user intended to install.

Signed-off-by: Mike Sul <mike.sul@foundries.io>