api: offline: Let user decide on target to install
Don't print the latest available/found target as "selected" in the
checkin local implementation. Instead, let a client/user decide which
if the found/available targets to consider as "selected" and/or install.
A decision which of the available targets to install is a use-case
specific.
Also, return zero targets if none of the tuf targets matches provided
content.

Signed-off-by: Mike Sul <mike.sul@foundries.io>