api: offline: Fix setting of package management type during CheckInLocal
Behavior of CheckInLocal depends on the package management type. If apps
use is enabled, only targets with all the required apps are returned.

In order to do so, this commit adds setPacmanType call inside
CheckInLocal as well, like we do when instantiating LocalLiteInstall,
and change CheckInLocal input type to LocalUpdateSource instead of 3
separate path, so we can test for docker_client_ptr as well.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>