tests: api: Add tests to unveil subtle issue
`AkliteClient::Installer` accepts a target json/object that
has a valid name however its content does not match the content of the
target json with the same name found in the tuf meta DB.
As a result, if the input target is hacked then an update/installation
is successful. As a matter of fact, the correct target is installed in
this case, the one that is found in the tuf meta db.
But, it is counter-intuitive, the installer should properly check and
reject such invalid/hacked targets before doing an update.

This commit contains just tests proving the issue exists - an
installation is successful while the installer should reject such an
input and even don't start an update.

Signed-off-by: Mike Sul <mike.sul@foundries.io>