api: offline: Don't gather apps to install
Since the offline update requires all target apps to be present in the
provided update source bundle then there is no reason to amend the
original TUF target stored in the TUF meta DB with apps to be installed.
If a target is found it means all its apps except shortlisted one should
be installed.

Signed-off-by: Mike Sul <mike.sul@foundries.io>