tests: offline: Unveil unavailable target installation issue
These two tests expose a subtle issue of the offline installation.
Current version allows creating an installer instance for a target
source content of which is not available.
It is not critical since the download fails in such case. However, it's
better to reject such updates earlier at the installer instantiation.
Otherwise it may confuse end users as well as can lead to accumulation
of ostree commits and/or app blobs if multiple unsuccessful update
attempts occurs.

Signed-off-by: Mike Sul <mike.sul@foundries.io>