api: offline: Make sure target to install is available
Make sure that content of the target specified to the installer is
actually available.
During the local checkin if update content of a given target is found
then extend its custom data with a new field that points to the content
location. Later, at the installer instantiation, verify whether the
specified target contains the reference to the locally available update
content.

Signed-off-by: Mike Sul <mike.sul@foundries.io>