api: Don't convert target type to filter targets
The current implementation does double conversion of target type to
filter out them by tags and a hardware ID. At first it convert the
aklite's tuf targets to the uptane's one, and then back to the tuf's
one.
This change gets rid of this double conversion as well as reduces
dependencies on libaktualizr and its uptane target type.

Signed-off-by: Mike Sul <mike.sul@foundries.io>