api: Distinguish checkin error types better
Add more error codes for the checkin results as well as corresponding
exit codes for the CLI API.
The goal is to distinguish between TUF repo update errors and the case
when a TUF repo is updated successfully, however there are no
TUF Targets that match a tag and/or hardware ID of a given device.

Also, improve log messages produced during the checking process so a
user can figure out a reason of a checkin failure.

Added corresponding unit tests.

Signed-off-by: Mike Sul <mike.sul@foundries.io>