api: Don't send events if offline update
There is no point in trying to send update events in the case of offline
update. Sending events in the download and install phases are preventing
by overriding the input config - setting the tls server url to an empty
string.
Unfortunately, it is not possible to apply the same trick to
the finalization phase since there is no update specific "installer"
instance.
This change makes the client not to send update events in the
finalization phase by checking the target custom field which is present
only in an offline update.

Signed-off-by: Mike Sul <mike.sul@foundries.io>