api: Use db metadata when offline pack has outdated TUF metadata
Offline checkin was never falling back to using the current stored
metadata. Now, if the offline checkin error is specifically related
to try to process outdated metadata, i.e., there is already a more
recent version in the local storage, we return an OK status (OkCached),
allowing an offline update attempt to proceed.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>