appengine: Don't require connectivity when starting app
The goal of the change is to address a need for connection with
Registry during app startup.
To achieve this, the following steps are taken:

- Images are loaded into the Docker store immediately after being
  pulled onto a device (into the Skopeo's image store).

- `docker compose pull` is invoked after the images have been loaded,
   it updates Docker engine's DB with digest references to the images.

As result, the subsequent call `docker compose up -d` does NOT require
a connection with Registry, and apps can be started when a device is
offline.

Also, the tests are adjusted to the changes in the `install` method
which cannot return `Result::ID::ImagePullFailure` error.

Signed-off-by: Mike Sul <mike.sul@foundries.io>