api: offline: Consider app shortlist when checking content
If an app shortlist is set in a factory config then the CI run that
fetches apps for offline update downloads only apps listed in the
shortlist. As a result, the offline bundle contains only apps listed in
the shortlist.
Therefore, the offline updater should require presence in a bundle only
those target apps that are in the shortlist
( the <target>.`custom.fetched-apps.shortlist` field).

Signed-off-by: Mike Sul <mike.sul@foundries.io>