offline: Don't check tags of the bundle targets
There is no reason to check if the bundle targets' tags and the device's
tag overlap/match as long as the updated device's TUF repo/meta are
valid and contains one or more of the bundle targets.

The device's and bundle's tag decides on the TUF metadata set to applied
on a device or downloaded to the bundle. If the device's TUF metadata
are updated without any error and checked/validated then that's fine to
apply one of the bundle targets as long as it is in a list of the
targets stored in the valid device TUF repo/DB and the device's hardware
ID matches the bundle target's one.

Signed-off-by: Mike Sul <mike.sul@foundries.io>