apps: Do not prune app manifest blobs
App manifest has at least one layer reference that points to the app
bundle and an optional layer reference that points to the blob
containing app layer metadata.

Also, the manifest can contain reference to a blob containing app layer
sizes. These references are stored in the `manifests` field of the app
manifest.

In addition to that the app manifest and the app image manifest can be
stored in the blob directory. If so, then make sure that they are nor
removed during app store pruning.

Signed-off-by: Mike Sul <mike.sul@foundries.io>