tests: improve reliability of events verification
Add a fsync operation inside the fake device gateway used for testing,
and also add a delay before reading the generated events file. The delay
is required because the sending of events inside aklite is not done
syncronously. A ReportQueue flush is forced on LiteClient destruction,
but we keep an instance active during the test procedure.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>