tests: Improve reliability of events verification
Add a fsync operation inside the fake device gateway used for testing,
a delay before reading the generated events file, and also a delay
before resetting the events list. The delay is required because the
sending of events inside aklite is not done synchronously. A ReportQueue
flush is forced on LiteClient destruction, but we keep an instance
active during the test procedure.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>