appengine: Remove app only from runtime part
Remove app components only from the runtime part, i.e. docker store and
compose project in the appengine->remove() call, and leave its content
in the OCI/blob store as it can be needed during rollback or if a given
app is listed in the `reset_apps`.
After successful update unused blobs are removed from the blob stores,
which guarantees proper app removal if it is not listed in `reset_apps`.

Signed-off-by: Mike Sul <mike.sul@foundries.io>