daemon: Do not skip sleep operation after successful install
After improvements in device gateway and aklite, the device's target
name at the server is updated right after installation is completed, no
need to go through a new check-in to do it.

This commit avoids an infinite loop without sleep if we hit an erroneous
situation where an install operation is called repeatedly, and always
return success.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>