Revert "app: Use `composectl` to check whether app is installed"
This reverts commit 37c0249d94ae1d399e170549acdb33bb5194a704.

The given functionality does not work if images hosted in docker.io are
specified in a docker compose project file in a full format.
For example, "docker.io/library/nginx:stable-alpine" or
"docker.io/homeassistant/home-assistant:stable".

The dockerd removes the "docker.io" or "docker.io/library" prefix when
stores such images in its store and as result returns the image URIs in
the shorten format in a response to "http://localhost/images/json?all=1"
request. For example, "nginx@sha256:<hash>" or
"homeassistant/home-assistant:<hash>".

Therefore, it is not possible to map images listed in the compose
project to images reported by dockerd unless the dockerd's URI shortening
rules are taken in the "isAppInstalled" logic.