liteclient: Truncate log of update complete event
Truncate a description of the update complete event if the completion is
not successful and it exceeds the maximum allowed size (2048).
The truncation is done by erasing the description middle part, so a user
can see the beginning of the app starting log and the end of it which is
supposed to contain a reason of failure.

Signed-off-by: Mike Sul <mike.sul@foundries.io>