api: extended: Select target with all apps when running an apps sync
Previous code was using the current target object, which is based on the
current running system, and may not contain all available apps.
This could lead to an issue where the apps sync update fails later on in
the online case, since equality between the latest available and
selected target is checked, and all apps must be included.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>