makefile: Fix make update command
The additional aktualizr-lite check command added in commit b628f99
(makefile: Add check operation on make update) introduced an issue
where the `make update` command stops before trying the aktualizr-lite
pull and install commands when there is an update to be performed, since
in that case, the check command returns an exit code != 0.

This commit ignores the return code of the check command, allowing the
`make update` command to always proceed to pull + install.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>