restorableappengine: Log app start
Print logs about an app start and successful completion.

Logs printed to stdout by processes/utilities spawned by aklite are
supressed since some of the utilities do odd/crazy stuff in the std
output which makes aklite logs messy. As drawback, it makes
successful execution some of utilities invisible, in particular
`docker compose up` which is an important to see in the logs.
This change prints logs before starting an app and just after
successfull execution, so it becomes visible when an app was started.
Note: if an app start failure occurs, then aklite will add the stderr
output of the command to the aklite's logs.

Signed-off-by: Mike Sul <mike.sul@foundries.io>