composeappmanager: Remove redundant app install check
There is no sense in doing such the primitive check whether an app is
installed in the compose app manager because both the restorable and
composeapp app engines do a proper installation check in the
"appengine->isRunning()" call.

Note: the composeapp app engine will include the installation check in
the following commit.

Signed-off-by: Mike Sul <mike.sul@foundries.io>