api: extended: Add explicit rollback operation
This new operation marks the current target as a failing target, and
proceeds to install the selected rollback target, which can be
automatically selected (the most recent fully installed target), or
explicitly defined by the caller of the function.
A rollback can be initiated after an installation is completed, but also
while finalization of system reboot is still pending.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>