api: Fix rollback to target that is not in TUF metadata
API code required the target to be present in TUF metadata in order to
allow its installation, even during a rollback operation. That leads to
rollback not working if the previous target was removed from TUF
metadata.
This commits fixes this issue by allowing the check to be skipped if
we are doing a rollback or apps sync update.
Considering the aklite daemon use, this was a regression introduced by
commit 7ed38b2 (main: Rewrite aklite daemon code using new
AkLiteClientExt methods), since before that the API code was not used by
the daemon.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>