api: extended: Fix apps sync when current target is not in TUF metadata
Commit 605b974 (api: extended: Select target with all apps when running
an apps sync) introduced a regression, where the current target is
required to be present in TUF metadata in order to an apps sync update
to be performed.
This change allows the target object derived from DB state to be used
instead of the one that was created from the TUF metadata when the later
is not available.
The fix implemented in 605b974 would not be required anymore since the
introduction of the commit previous to this one (api: Fix rollback to
target that is not in TUF metadata). But instead of just reverting it,
the current commit prioritizes the use of the target derived from TUF
metadata for consistency with the typical use scenario.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>