api: Prevent downgrade operations to be done automatically
When the latest available target has a version lower than the target
that is currently running, we may or not accept it as a valid update
to be performed.

Before this commit, aktualizr-lite would automatically select such
target and proceed with the update operation. After this commit, the
target is not longer selected by default as a valid update target.

It is possible to revert the behavior to the original one by defining
AUTO_DOWNGRADE at build time.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>